Initial Caching Interface - wip
