+++ web/concrete/config/app.php
@@ -333,6 +333,7 @@ return array(
         "/ccm/system/panels/multilingual"                                               => array('\Concrete\Controller\Panel\Multilingual::view'),
         "/ccm/system/panels/page"                                                       => array('\Concrete\Controller\Panel\Page::view'),
         "/ccm/system/panels/page/attributes"                                            => array('\Concrete\Controller\Panel\Page\Attributes::view'),
+        "/ccm/system/panels/page/caching"                                               => array('\Concrete\Controller\Panel\Page\Caching::view'),
         "/ccm/system/panels/page/check_in"                                              => array('\Concrete\Controller\Panel\Page\CheckIn::__construct'),
         "/ccm/system/panels/page/check_in/submit"                                       => array('\Concrete\Controller\Panel\Page\CheckIn::submit'),
         "/ccm/system/panels/page/design"                                                => array('\Concrete\Controller\Panel\Page\Design::view'),
@@ -364,6 +365,7 @@ return array(
         "/ccm/system/panels/details/page/attributes/add_attribute"                      => array('\Concrete\Controller\Panel\Detail\Page\Attributes::add_attribute'),
         "/ccm/system/panels/details/page/attributes/submit"                             => array('\Concrete\Controller\Panel\Detail\Page\Attributes::submit'),
         "/ccm/system/panels/details/page/caching"                                       => array('\Concrete\Controller\Panel\Detail\Page\Caching::view'),
+        "/ccm/system/panels/details/page/caching/preview"                               => array('\Concrete\Controller\Panel\Detail\Page\Caching::preview'),
         "/ccm/system/panels/details/page/caching/purge"                                 => array('\Concrete\Controller\Panel\Detail\Page\Caching::purge'),
         "/ccm/system/panels/details/page/caching/submit"                                => array('\Concrete\Controller\Panel\Detail\Page\Caching::submit'),
         "/ccm/system/panels/details/page/composer"                                      => array('\Concrete\Controller\Panel\Detail\Page\Composer::view'),
