+++ web/concrete/elements/block_header_view.php
@@ -41,6 +41,13 @@ if (
     ));
 }
 
+if(\Config::get('concrete.cache.preview')) {
+    if(!$b->useBlockCache()) {
+        echo "<div class='ccm-cache-status' style='outline: solid 1px red'>
+                <span class='label label-danger' style='position:absolute; z-index:3000;'>{$b->instance->getBlockTypeName()} ({$b->getBlockID()})</span>";
+    }
+}
+
 if ($showMenu) {
     $arHandle = $a->getAreaHandle();
